pyproject.toml
[project]
name = "antenna"
version = "0.1.0"
description = "Local-first RSS/Atom/JSON Feed reader for email and MCP."
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Todd Deshane" }]
license = "MIT"
license-files = ["LICENSE"]
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Development Status :: 3 - Alpha",
]
dependencies = [
    "feedparser>=6.0.11",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "mcp>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/toddllm/antenna"
Issues = "https://github.com/toddllm/antenna/issues"

[project.scripts]
antenna = "antenna.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["antenna*"]

[tool.setuptools.package-data]
antenna = ["templates/*.html"]
